+++ src/components/search/Filters/Panels/AvailabilityFilterPanel.tsx
