+++ src/components/search/Filters/Panels/MinFilterPanel.tsx
