+++ src/components/search/Filters/Panels/SectionTypeFilterPanel.tsx
